requirements.txt
gradio>=4.0.0
openai-whisper>=20231117
torch>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
gunicorn>=21.2.0

